Remove admin statement parser / analyzer deprecated code
