--- fe/fe-core/src/main/java/com/starrocks/analysis/AdminRepairTableStmt.java
@@ -1,75 +1,0 @@
-// This file is made available under Elastic License 2.0.
-// This file is based on code available under the Apache license here:
-//   https://github.com/apache/incubator-doris/blob/master/fe/fe-core/src/main/java/org/apache/doris/analysis/AdminRepairTableStmt.java
-
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-package com.starrocks.analysis;
-
-import com.google.common.collect.Lists;
-import com.starrocks.sql.ast.AstVisitor;
-
-import java.util.List;
-
-//  ADMIN REPAIR TABLE table_name partitions;
-public class AdminRepairTableStmt extends DdlStmt {
-
-    private TableRef tblRef;
-    private List<String> partitions = Lists.newArrayList();
-
-    private long timeoutS = 0;
-
-    public AdminRepairTableStmt(TableRef tblRef) {
-        this.tblRef = tblRef;
-    }
-    public void setDbName(String dbName) {
-        this.tblRef.getName().setDb(dbName);
-    }
-    public void setPartitions(PartitionNames partitionNames) {
-        this.partitions.addAll(partitionNames.getPartitionNames());
-    }
-    public void setTimeoutSec(long timeoutS) {
-        this.timeoutS = timeoutS;
-    }
-
-    @Override
-    public <R, C> R accept(AstVisitor<R, C> visitor, C context) {
-        return visitor.visitAdminRepairTableStatement(this, context);
-    }
-    @Override
-    public boolean isSupportNewPlanner() {
-        return true;
-    }
-
-    public String getDbName() {
-        return tblRef.getName().getDb();
-    }
-
-    public String getTblName() {
-        return tblRef.getName().getTbl();
-    }
-
-    public PartitionNames getPartitionNames() {return tblRef.getPartitionNames();}
-    public List<String> getPartitions() {
-        return partitions;
-    }
-
-    public long getTimeoutS() {
-        return timeoutS;
-    }
-}
