+++ fe/fe-core/src/main/java/com/starrocks/catalog/MetadataViewer.java
@@ -23,8 +23,6 @@ package com.starrocks.catalog;
 
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
-import com.starrocks.analysis.AdminShowReplicaDistributionStmt;
-import com.starrocks.analysis.AdminShowReplicaStatusStmt;
 import com.starrocks.analysis.BinaryPredicate.Operator;
 import com.starrocks.analysis.PartitionNames;
 import com.starrocks.catalog.MaterializedIndex.IndexExtState;
@@ -33,6 +31,8 @@ import com.starrocks.catalog.Table.TableType;
 import com.starrocks.common.DdlException;
 import com.starrocks.common.FeConstants;
 import com.starrocks.server.GlobalStateMgr;
+import com.starrocks.sql.ast.AdminShowReplicaDistributionStmt;
+import com.starrocks.sql.ast.AdminShowReplicaStatusStmt;
 import com.starrocks.system.Backend;
 import com.starrocks.system.SystemInfoService;
