+++ fe/fe-core/src/main/java/com/starrocks/execution/AdminSetReplicaStatusExecutor.java
@@ -2,11 +2,11 @@
 
 package com.starrocks.execution;
 
-import com.starrocks.analysis.AdminSetReplicaStatusStmt;
 import com.starrocks.analysis.StatementBase;
 import com.starrocks.common.DdlException;
 import com.starrocks.qe.ConnectContext;
 import com.starrocks.qe.ShowResultSet;
+import com.starrocks.sql.ast.AdminSetReplicaStatusStmt;
 
 public class AdminSetReplicaStatusExecutor implements DataDefinitionExecutor {
