+++ fe/fe-core/src/main/java/com/starrocks/sql/ast/AdminSetConfigStmt.java
@@ -1,0 +1,52 @@
+// This file is licensed under the Elastic License 2.0. Copyright 2021-present, StarRocks Inc.
+
+package com.starrocks.sql.ast;
+
+import com.google.common.collect.Maps;
+import com.starrocks.analysis.DdlStmt;
+import com.starrocks.analysis.RedirectStatus;
+
+import java.util.Map;
+
+// admin set frontend config ("key" = "value");
+public class AdminSetConfigStmt extends DdlStmt {
+
+    public enum ConfigType {
+        FRONTEND,
+        BACKEND
+    }
+
+    private final ConfigType type;
+    private Map<String, String> configs;
+
+    public AdminSetConfigStmt(ConfigType type, Map<String, String> configs) {
+        this.type = type;
+        this.configs = configs;
+        if (this.configs == null) {
+            this.configs = Maps.newHashMap();
+        }
+    }
+
+    public ConfigType getType() {
+        return type;
+    }
+
+    public Map<String, String> getConfigs() {
+        return configs;
+    }
+
+    @Override
+    public RedirectStatus getRedirectStatus() {
+        return RedirectStatus.NO_FORWARD;
+    }
+
+    @Override
+    public <R, C> R accept(AstVisitor<R, C> visitor, C context) {
+        return visitor.visitAdminSetConfigStatement(this, context);
+    }
+
+    @Override
+    public boolean isSupportNewPlanner() {
+        return true;
+    }
+}
