+++ fe/fe-core/src/main/java/com/starrocks/sql/ast/AdminShowConfigStmt.java
@@ -1,32 +1,13 @@
-// This file is made available under Elastic License 2.0.
-// This file is based on code available under the Apache license here:
-//   https://github.com/apache/incubator-doris/blob/master/fe/fe-core/src/main/java/org/apache/doris/analysis/AdminShowConfigStmt.java
+// This file is licensed under the Elastic License 2.0. Copyright 2021-present, StarRocks Inc.
 
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-package com.starrocks.analysis;
+package com.starrocks.sql.ast;
 
 import com.google.common.collect.ImmutableList;
-import com.starrocks.analysis.AdminSetConfigStmt.ConfigType;
+import com.starrocks.analysis.ShowStmt;
 import com.starrocks.catalog.Column;
 import com.starrocks.catalog.ScalarType;
 import com.starrocks.qe.ShowResultSetMetaData;
-import com.starrocks.sql.ast.AstVisitor;
+import com.starrocks.sql.ast.AdminSetConfigStmt.ConfigType;
 
 // admin show frontend config;
 public class AdminShowConfigStmt extends ShowStmt {
@@ -34,9 +15,9 @@ public class AdminShowConfigStmt extends ShowStmt {
             new ImmutableList.Builder<String>().add("Key").add("AliasNames").add(
                     "Value").add("Type").add("IsMutable").add("Comment").build();
 
-    private ConfigType type;
+    private final ConfigType type;
 
-    private String pattern;
+    private final String pattern;
 
     public AdminShowConfigStmt(ConfigType type, String pattern) {
         this.type = type;
