+++ fe/fe-core/src/test/java/com/starrocks/catalog/AdminStmtTest.java
@@ -22,7 +22,7 @@
 package com.starrocks.catalog;
 
 import com.google.common.collect.Lists;
-import com.starrocks.analysis.AdminSetReplicaStatusStmt;
+import com.starrocks.sql.ast.AdminSetReplicaStatusStmt;
 import com.starrocks.analysis.CreateDbStmt;
 import com.starrocks.analysis.CreateTableStmt;
 import com.starrocks.catalog.MaterializedIndex.IndexExtState;
