+++ fe/fe-core/src/test/java/com/starrocks/sql/analyzer/AdminSetTest.java
@@ -1,7 +1,7 @@
 // This file is licensed under the Elastic License 2.0. Copyright 2021-present, StarRocks Inc.
 package com.starrocks.sql.analyzer;
 
-import com.starrocks.analysis.AdminSetReplicaStatusStmt;
+import com.starrocks.sql.ast.AdminSetReplicaStatusStmt;
 import com.starrocks.utframe.UtFrameUtils;
 import org.junit.Assert;
 import org.junit.BeforeClass;
